Builder reworked into util method
