+++ src/test/kotlin/com/exactpro/th2/http/server/testimpl/TestServerManager.kt
@@ -16,12 +16,14 @@ package com.exactpro.th2.http.server.testimpl
 
 import com.exactpro.th2.common.grpc.Direction
 import com.exactpro.th2.common.grpc.EventID
+import com.exactpro.th2.common.grpc.MessageGroup
 import com.exactpro.th2.common.grpc.RawMessage
 import com.exactpro.th2.common.message.addField
 import com.exactpro.th2.common.message.message
+import com.exactpro.th2.common.message.plusAssign
 import com.exactpro.th2.http.client.HttpClient
 import com.exactpro.th2.http.server.HttpServer
-import com.exactpro.th2.http.server.util.ResponseBuilder
+import com.exactpro.th2.http.server.util.toResponse
 import com.google.protobuf.ByteString
 import mu.KotlinLogging
 import org.junit.jupiter.api.Assertions
@@ -67,7 +69,10 @@ open class TestServerManager(private val https: Boolean = false, socketDelayChec
             metadata = metadataBuilder.putProperties("contentType", "application").build()
         }.build()
 
-        ResponseBuilder().setHead(responseMessage).setBody(bodyMessage).build()
+        MessageGroup.newBuilder().apply {
+            this += responseMessage
+            this += bodyMessage
+        }.toResponse()
     }
 
     fun start() {
